tidyverse-package {tidyverse}	R Documentation

tidyverse: Easily Install and Load the 'Tidyverse'

Description

The 'tidyverse' is a set of packages that work in harmony because they share common data representations and 'API' design. This package is designed to make it easy to install and load multiple 'tidyverse' packages in a single step. Learn more about the 'tidyverse' at <https://tidyverse.org>.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

http://tidyverse.tidyverse.org

https://github.com/tidyverse/tidyverse

Report bugs at https://github.com/tidyverse/tidyverse/issues

[Package tidyverse version 1.3.0 Index]