tidyverse_sitrep {tidyverse}	R Documentation

Get a situation report on the tidyverse

Description

This function gives a quick overview of the versions of R and RStudio as well as all tidyverse packages. It's primarily designed to help you get a quick idea of what's going on when you're helping someone else debug a problem.

Usage

tidyverse_sitrep()

[Package tidyverse version 1.3.0 Index]