tidyverse_update {tidyverse}	R Documentation

Update tidyverse packages

Description

This will check to see if all tidyverse packages (and optionally, their dependencies) are up-to-date, and will install after an interactive confirmation.

Usage

tidyverse_update(recursive = FALSE, repos = getOption("repos"))

Arguments

recursive
If TRUE, will also list all dependencies of tidyverse packages.

repos
The repositories to use to check for updates. Defaults to getOption("repos").

Examples

## Not run: 
tidyverse_update()

## End(Not run)

[Package tidyverse version 2.0.0 Index]