tidyverse: Easily Install and Load the 'Tidyverse'
Source: R/tidyverse.R

The 'tidyverse' is a set of packages that work in harmony because they share common data representations and 'API' design. This package is designed to make it easy to install and load multiple 'tidyverse' packages in a single step. Learn more about the 'tidyverse' at <https://tidyverse.org>.

See also

Useful links:

http://tidyverse.tidyverse.org

https://github.com/tidyverse/tidyverse

Report bugs at https://github.com/tidyverse/tidyverse/issues