tidyverse_conflicts {tidyverse}	R Documentation

Conflicts between the tidyverse and other packages

Description

This function lists all the conflicts between packages in the tidyverse and other packages that you have loaded.

Usage

tidyverse_conflicts(only = NULL)

Arguments

only
Set this to a character vector to restrict to conflicts only with these packages.

Details

There are four conflicts that are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

tidyverse_conflicts()

[Package tidyverse version 2.0.0 Index]