Conflicts between the tidyverse and other packages
Source: R/conflicts.R

This function lists all the conflicts between packages in the tidyverse and other packages that you have loaded.

tidyverse_conflicts()

Details

There are four conflicts that are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

tidyverse_conflicts()

#> ── Conflicts ───────────────────────────────────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter()                   masks stats::filter()
#> ✖ purrr::is_null()                  masks testthat::is_null()
#> ✖ dplyr::lag()                      masks stats::lag()
#> ✖ tidyverse::library.dynam.unload() masks base::library.dynam.unload()
#> ✖ dplyr::matches()                  masks testthat::matches()
#> ✖ tidyverse::package_version()      masks base::package_version()
#> ✖ devtools_shims::system.file()     masks tidyverse::system.file(), base::system.file()
#> ✖ dplyr::vars()                     masks ggplot2::vars()